requirements.txt
Django==3.2
Pillow
mysqlclient
django-allauth
pymysql